ULON-447, ULON-449 (#997)
* ULON-351

설비용량 소수점 표시

* ULON-329

시작일 종료일 각각 15일 씩 빼고 더하기

* ULON-331

다운로드 리스트에 넣기

* ULON-330

자산운용사가 승인완료에서 증인중으로 변경가능하도록 수정

* ULON-345

어드민계정으로 로그인시 수정가능

* 관리자일경우 버튼 보여주기

* ULON-345

잘못 수정된 부분 다시 수정

* ULON-339

role이 1일 경우 메뉴 전체 오픈

* ULON-215

테이블 기반 대시보드 1차 작업

* 수정

* ULON-341

권한 등급에 따라 하단 탭영역 표시

* ULON-345

* 오류 수정

* ULON-350

station_id 공통으로 변경

* KPX-LOCAL 세팅

* BNDT-19

소스 품질 개선

* ULON-341

알람 이력 파일 올리기 수정

* ULON-341

수정

* ULON-341

수정

* ULON-352

이상분석 시간 계산 추가

* ULON-284

권한에 따른 제약

* ULON-358

수정

* 수정

* 코드값 변경

* 코드 값으로 표시 되는 문제 수정

* ULON-337 ULON-352

설비용량 계산오류 수정

* 콘솔로그 삭제

* 운영이행시 테이블 대시보드 삭제

운영이행시 테이블 대시보드 삭제

* KPX 수정 사항

* KPX 수정 사항 적용

KPX 수정 사항 적용

* ULON-328

리프레시를 해도 검색조건 유지하도록 작업

* 사용 안하는 클래스 삭제

* KPX 수정사항 적용

* Auto stash before merge of "encoredModify" and "upstream/iDERMS-SaaS"

* 일부 오류 때문에 수정

* 잘못 걸린 조건 수정

* KPX 시에 전일 데이터 기준일 오늘로 변경

* KPX API 설정 변경

* 그룹대시보드 소스 전면 수정 && 테이블 대시 보드

* 머지

* 테이블 대시보드

* 주석

* 수정

* replace오류 수정

* ULON-311

인증서 리스트 정렬

* 대시 보드 차트 일부 수정

* BNDT-21

아이디 찾기 비밀번호 찾기 추가 세션 타임아웃 15분 세팅

* ULON-367

대시보드 툴팁개선

* Auto stash before merge of "encoredModify" and "upstream/iDERMS-SaaS"

* 오류 수정

* ULON-372

로그인 오류 메시지 수정

* 수정

* 수정

* Auto stash before merge of "encoredModify" and "upstream/iDERMS-SaaS"

* 툴팁 수정

* 단위 수정

* 송신시간 표기 오류 수정

* 단위 수정

* 대시 보드 목표출력 단위 수정

* 툴팁 단위 오류 수정

* 알람 코드 추가

* 수정

* 프로퍼티 세팅 encored로 다시 변경 kpx 그룹대시보드 테이블 수정사항 적용

* 상태 체크 일부 동작 수정

* 일부 수정

* 텍스트 오류 수정

* wpsolar Logo

wpsolar Logo 세팅

* 파비콘 svg -> ico로 변경

* svg 파비콘 파일 삭제

* ULON-328

페이지 검색 기록 유지 세팅

* ULON-325

error_calculator api 보낼시에 did세팅해서 보내도록 수정

* 머지 후 다시 수정

* 수정

* 그룹대시보드 일부수정

대시보드 변경안되는 부분 수정

* 수정

* 대시보드 일부 수정

대시보드에서 사용하는 energy/sites && forecasting/site v2로 수정

* ULON-326 그외

ULON-326 
직집입력 맨위로 올리기

그룹대시보드 일부 정보 오류 수정

* 사이트 세팅 오류 수정

* ULON-383

SPC명 추가

* 중개거래 대시보드 스크립트 공통화

* 중개거래 대시보드 수정적용

* ULON-225 검색 필터 오류 수정

* 합계 수정

* 소스 다시 원복

* 디스플레이 텍스트 오류 수정

* /weather/site V2로 변경

* 오류 수정

* 잘못 머지 된 파일 다시 수정

* ULON-387

* ULON-388

목록버튼 오동작 수정

* 테이블 데이터 수정

* ULON-393

알람코드 세팅

* dropdown 초기화 && 수정시 세팅

* ULON-394

* ULON-392

페이징 버튼 작동안함

* ULON-393

* 알람코드 단건 선택으로 변경

* Auto stash before merge of "encoredModify" and "upstream/iDERMS-SaaS"

* 페이지 오류 수정

* ULON-393, ULON-404, ULON-402

* 수정사항 적용

* ULON-408

날씨 정렬 기준 불명확하므로 정렬 제거

* 테스트

* 버튼 클래스 수정

* 오류 수정

* ULON-389

시리즈 숨길시에 발생하는 y축 타이틀이 x축 카테고리와 겹치는 문제 수정

* 테스트용 코드 삭제

* 테스트 코드 삭제

* ULON-391

* 사업소 관리 기상 그리드 X,Y 추가

* 수정

* ULON-423

날씨 데이터가 존재하지않을때 표시 제외
날씨데이터 조회 방식 한번에 조회하도록 수정

* ULON-418

엑셀 업로드 오류 발생시 fileinput 초기화

* ULON-424

대시보드에서 이동시 기본세팅 1시간 단위로 조회

* ULON-417

SPC 선택시 기본정보 입력되있던 기준으로 세팅

* ULON-432

정렬 제대로 안되는 문제 수정

* 사이트명 검색오류 수정

* ULON-417

SPC 선택시 기존 정보 세팅해주기

* ULON-435

리스트 datatables로 변경

* 작업 보고서 일부 수정

* ULON-420

메세지를 삭제문구로 쓰던것을 '삭제'로 픽스

code_set code 중에 code가 없는 항목들이 존재하는것을 확인
code가 없을경우 api 호출시 제외하도록 수정

* ULON-440

타입 원상태로 변경

* ULON-439

kWh 단위로 변경

* 수정

* ULON-444, ULON-445, 

ULON-444
기상그리드 입력 가능 길이 4자리

ULON-445
수정시에 지역 수정안되는 문제 수정

* ULON-427

작업보고서 항목 추가

* ULON-447

기상그리드 위경도 입력시 자동세팅하도록 적용

* 테이블 헤더 넓이 문제 수정

* ULON-477

기상그리드 계산 버튼 적용

* ULON-449

ESS 값 의미 없게 쓰는 부분  쓰는 것으로 수정

* 알람 히스토리 오류 수정

* 사업소 관리

사이트명 공백 trim한후 조건 체크하게 추가 
기상그리드 숫자키만 입력받게 변경

* 알람 세팅 삭제 오류 수정

* 수정

+++ energyServicePortal_ewp_adv/src/main/webapp/WEB-INF/jsp/esp/setting/alarmSetting.jsp
@@ -174,12 +174,15 @@
 		}).columns.adjust().draw();
 
 		$("#comDeleteBtn").click(function(){
+			$("#comDeleteModal").modal('hide');
+			$("#confirmTitle").val('');
+
 			const alarmTable = $('#alarmTable').DataTable();
 			const checkedArray = document.querySelectorAll('[name="table_checkbox"]:checked');
-			let modal = $("#comDeleteModal");
-			let modalBody = $("#deleteConfirmModal .modal-body");
-			let urls = new Array();
-			let deferreds = new Array();
+			const urls = new Array();
+			const deferreds = new Array();
+			const deferreds2 = new Array();
+			const deferreds3 = new Array();
 
 			if ($('#regist').prop('disabled')) { //조회한 데이터 일경우
 				checkedArray.forEach(chk => {
@@ -229,7 +232,7 @@
 
 					selectURL.forEach(function (url) {
 						let deferred = $.Deferred();
-						deferreds.push(deferred);
+						deferreds2.push(deferred);
 
 						$.ajax(url).done(function (data) {
 							data['url'] = url['url'];
@@ -241,52 +244,53 @@
 						});
 					});
 
-					$.when.apply($, deferreds).then(function () {
+					$.when.apply($, deferreds2).then(function () {
 						let deleteURL = new Array();
+						setIdArray = new Array();
+
 						for (let index = 0; index < arguments.length; index++) {
 							let data = arguments[index];
 
 							if (isEmpty(data.data)) {
-								setIdArray = new Array();
-
 								let url = data.url.split('/');
 								setIdArray.push(url[url.length - 2]);
 							}
 						}
 
-						//codeSet 삭제
-						setIdArray.forEach(setId => {
-							deleteURL.push({
-								url: apiHost + '/alarms/code_sets/' + setId,
-								type: 'delete',
-								dataType: 'json'
+						if (isEmpty(setIdArray)) {
+							//codeSet 삭제
+							setIdArray.forEach(setId => {
+								deleteURL.push({
+									url: apiHost + '/alarms/code_sets/' + setId,
+									type: 'delete',
+									dataType: 'json'
+								});
 							});
-						});
 
-						deleteURL.forEach(function (url) {
-							let deferred = $.Deferred();
-							deferreds.push(deferred);
+							deleteURL.forEach(function (url) {
+								let deferred = $.Deferred();
+								deferreds3.push(deferred);
 
-							$.ajax(url).done(function (data) {
-								data['url'] = url['url'];
-								(function (deferred) {
-									return deferred.resolve(data);
-								})(deferred);
-							}).fail(function (error) {
-								console.log(error);
+								$.ajax(url).done(function (data) {
+									data['url'] = url['url'];
+									(function (deferred) {
+										return deferred.resolve(data);
+									})(deferred);
+								}).fail(function (error) {
+									console.log(error);
+								});
 							});
-						});
 
-						$.when.apply($, deferreds).then(function () {
-							modalBody.addClass("hidden");
-							$("#comDeleteSuccessMsg").text("알람 메세지가 삭제 되었습니다.").removeClass("hidden");
+							$.when.apply($, deferreds3).then(function () {
+								schAlarmList();
+								errorMsg('알람 메세지가 삭제 되었습니다.');
+								return false;
+							});
+						} else {
 							schAlarmList();
-							setTimeout(function(){
-								modal.modal("hide");
-							}, 1000);
-
+							errorMsg('알람 메세지가 삭제 되었습니다.');
 							return false;
-						});
+						}
 					});
 				});
 			} else { //엑셀 업로드 일경우
@@ -687,6 +691,7 @@
 				alert('등록 되었습니다.');
 				alarmTable.clear().draw();
 
+				$('#excelUploadBtn').val('');
 				$('#regist').prop('disabled', true);
 			});
 		}).fail(function (jqXHR, textStatus, errorThrown) {
@@ -725,12 +730,12 @@
 	 * @param msg
 	 */
 	const errorMsg = msg => {
-		$('#excelUploadBtn').val(''); //에러 발생시 fileInput 초기화
+		$('#excelUploadBtn').val('');
 
-		$("#warningMsg").text(msg);
-		$("#warningModal").modal("show");
+		$("#errMsg").text(msg);
+		$("#errorModal").modal("show");
 		setTimeout(function(){
-			$("#warningModal").modal("hide");
+			$("#errorModal").modal("hide");
 		}, 1800);
 	}
 </script>
